1-5-1008 2nd commit

+++ 1-5-1008.c
@@ -3,6 +3,6 @@
 int main(void) {
     int A, B;
     scanf("%d %d", &A, &B);
-    printf("%f\n", A / B);
+    printf("%lf\n", (double)A / B);
     return 0;
 }
